gradio/requirements.txt
gradio>=4.0.0
transformers>=4.37.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
qwen-vl-utils
